requirements.txt
fastapi>=0.90.0
uvicorn[standard]>=0.20.0
requests>=2.28.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
google-cloud-bigquery>=3.10.0
google-cloud-secret-manager>=2.11.0   # optional, if using secrets
python-dotenv>=1.0.0                  # optional, for local env files
pydantic>=1.10.0
